move bech32 package to internal
